decouple comment list and responsiveness fix
