+++ portfolio/src/main/webapp/index.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html lang=en><head><meta charset=utf-8><meta http-equiv=X-UA-Compatible content="IE=edge"><meta name=viewport content="width=device-width,initial-scale=1"><link rel=icon href=/favicon.ico><link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900" rel=stylesheet><link href=https://cdn.jsdelivr.net/npm/@mdi/font@4.x/css/materialdesignicons.min.css rel=stylesheet><title>Nicolas Capparelli: Portfolio</title><link href=/css/app.8a5727cf.css rel=preload as=style><link href=/css/chunk-vendors.dd78ee4f.css rel=preload as=style><link href=/js/app.92da4305.js rel=preload as=script><link href=/js/chunk-vendors.025af9e1.js rel=preload as=script><link href=/css/chunk-vendors.dd78ee4f.css rel=stylesheet><link href=/css/app.8a5727cf.css rel=stylesheet></head><body><noscript><strong>We're sorry but portfolio doesn't work properly without JavaScript enabled. Please enable it to continue.</strong></noscript><div id=app></div><script src=/js/chunk-vendors.025af9e1.js></script><script src=/js/app.92da4305.js></script></body></html>
+<!DOCTYPE html><html lang=en><head><meta charset=utf-8><meta http-equiv=X-UA-Compatible content="IE=edge"><meta name=viewport content="width=device-width,initial-scale=1"><link rel=icon href=/favicon.ico><link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900" rel=stylesheet><link href=https://cdn.jsdelivr.net/npm/@mdi/font@4.x/css/materialdesignicons.min.css rel=stylesheet><title>Nicolas Capparelli: Portfolio</title><link href=/css/app.7bbe5627.css rel=preload as=style><link href=/css/chunk-vendors.dd78ee4f.css rel=preload as=style><link href=/js/app.4d980acc.js rel=preload as=script><link href=/js/chunk-vendors.025af9e1.js rel=preload as=script><link href=/css/chunk-vendors.dd78ee4f.css rel=stylesheet><link href=/css/app.7bbe5627.css rel=stylesheet></head><body><noscript><strong>We're sorry but portfolio doesn't work properly without JavaScript enabled. Please enable it to continue.</strong></noscript><div id=app></div><script src=/js/chunk-vendors.025af9e1.js></script><script src=/js/app.4d980acc.js></script></body></html>
