--- portfolio/src/main/webapp/js/app.92da4305.js
@@ -1,2 +1,0 @@
-(function(t){function e(e){for(var i,s,o=e[0],c=e[1],l=e[2],m=0,p=[];m<o.length;m++)s=o[m],Object.prototype.hasOwnProperty.call(a,s)&&a[s]&&p.push(a[s][0]),a[s]=0;for(i in c)Object.prototype.hasOwnProperty.call(c,i)&&(t[i]=c[i]);u&&u(e);while(p.length)p.shift()();return r.push.apply(r,l||[]),n()}function n(){for(var t,e=0;e<r.length;e++){for(var n=r[e],i=!0,o=1;o<n.length;o++){var c=n[o];0!==a[c]&&(i=!1)}i&&(r.splice(e--,1),t=s(s.s=n[0]))}return t}var i={},a={app:0},r=[];function s(e){if(i[e])return i[e].exports;var n=i[e]={i:e,l:!1,exports:{}};return t[e].call(n.exports,n,n.exports,s),n.l=!0,n.exports}s.m=t,s.c=i,s.d=function(t,e,n){s.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:n})},s.r=function(t){"undefined"!==typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},s.t=function(t,e){if(1&e&&(t=s(t)),8&e)return t;if(4&e&&"object"===typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(s.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var i in t)s.d(n,i,function(e){return t[e]}.bind(null,i));return n},s.n=function(t){var e=t&&t.__esModule?function(){return t["default"]}:function(){return t};return s.d(e,"a",e),e},s.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},s.p="/";var o=window["webpackJsonp"]=window["webpackJsonp"]||[],c=o.push.bind(o);o.push=e,o=o.slice();for(var l=0;l<o.length;l++)e(o[l]);var u=c;r.push([0,"chunk-vendors"]),n()})({0:function(t,e,n){t.exports=n("56d7")},"034f":function(t,e,n){"use strict";var i=n("85ec"),a=n.n(i);a.a},"05c0":function(t,e,n){"use strict";var i=n("b2d5"),a=n.n(i);a.a},1182:function(t,e,n){},1395:function(t,e,n){},"15f2":function(t,e,n){},"1a74":function(t,e,n){},"34ba":function(t,e,n){"use strict";var i=n("1182"),a=n.n(i);a.a},4276:function(t,e,n){},"52a4":function(t,e,n){"use strict";var i=n("a845"),a=n.n(i);a.a},"56d7":function(t,e,n){"use strict";n.r(e);n("e260"),n("e6cf"),n("cca6"),n("a79d");var i=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"app"}},[n("v-app",[n("link",{attrs:{href:"https://fonts.googleapis.com/css?family=Roboto:400,500,700,400italic|Material+Icons",rel:"stylesheet"}}),n("NavBar"),n("v-content",[n("router-view",{key:t.$route.fullPath})],1)],1)],1)},a=[],r=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"nav-container"}},[n("v-app-bar",{attrs:{id:"toolbar",app:"",fixed:"",color:"#343333",elevation:"2",height:"auto"}},[n("v-list",{attrs:{id:"toolbar-list"}},[n("v-list-item",{staticStyle:{padding:"0"},attrs:{"two-line":""}},[n("v-list-item-avatar",[n("v-img",{attrs:{src:"/assets/me.jpeg"}})],1),n("v-list-item-content",[n("v-list-item-title",{staticStyle:{"font-weight":"bold","font-size":"18px"}},[t._v("Nicolas Capparelli")]),n("v-list-item-subtitle",{staticStyle:{"font-size":"14px"}},[t._v("Software Engineer")])],1)],1)],1),n("v-spacer"),n("v-app-bar-nav-icon",{on:{click:function(e){t.drawer=!t.drawer}}})],1),n("v-navigation-drawer",{attrs:{temporary:"",fixed:"",right:""},model:{value:t.drawer,callback:function(e){t.drawer=e},expression:"drawer"}},[n("v-list-item",[n("v-list-item-content",[n("v-list-item-title",[t._v(t._s(t.currentPage))])],1)],1),t.isUserLoggedIn?n("v-list-item",[n("v-list-item-content",[n("v-list-item-title",[t._v(t._s(t.userEmail))])],1)],1):t._e(),n("v-divider"),n("v-list",[t._l(t.pageList,(function(e){return n("v-list-item",{key:e.name,attrs:{link:""},on:{click:function(n){return t.navigateTo(e.name)}}},[n("v-list-item-icon",[n("v-icon",[t._v(t._s(e.icon))])],1),n("v-list-item-content",[n("v-list-item-title",[t._v(t._s(e.name))])],1)],1)})),t.isUserLoggedIn?n("v-list-item",{attrs:{link:""},on:{click:function(e){return t.logout()}}},[n("v-list-item-icon",[n("v-icon",[t._v("mdi-account")])],1),n("v-list-item-content",[n("v-list-item-title",[t._v("Logout")])],1)],1):t._e()],2)],1)],1)},s=[],o=(n("d3b7"),n("96cf"),n("1da1")),c="https://8080-b83fc153-d2cf-481d-a321-9342cdf80f21.us-east1.cloudshell.dev",l={textPrimary:"#DBDBDB"},u={name:"Template",created:function(){var t=this;this.$root.$on("pchange",(function(e){t.currentPage=e})),this.$root.$on("authCheck",(function(e){t.isUserLoggedIn=e.isLoggedIn,t.userEmail=e.email}))},methods:{navigateTo:function(t){this.currentPage=t,this.$router.push({name:t})},logout:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(){var e,n;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return t.next=2,fetch(c+"/authenticate");case 2:if(e=t.sent,!e.ok){t.next=10;break}return t.next=6,e.json();case 6:n=t.sent,location.href=n.authURL,t.next=11;break;case 10:alert("There was an issue processing your request, please try again later");case 11:case"end":return t.stop()}}),t)})));function e(){return t.apply(this,arguments)}return e}()},data:function(){return{drawer:!1,currentPage:"Home",pageList:[{name:"Home",icon:"mdi-home"},{name:"Resume",icon:"mdi-file"},{name:"Projects",icon:"mdi-code-greater-than-or-equal"},{name:"About",icon:"mdi-information-outline"},{name:"Connect",icon:"mdi-email"}]}}},m=u,p=(n("34ba"),n("2877")),d=Object(p["a"])(m,r,s,!1,null,null,null),h=d.exports,g={name:"App",components:{NavBar:h}},v=g,f=(n("034f"),Object(p["a"])(v,i,a,!1,null,null,null)),b=f.exports,y=n("2b0e"),w=n("67b0"),_=n("ce5b"),k=n.n(_);n("bf40");y["default"].use(k.a);var x,C,S={theme:{dark:!0}},j=new k.a(S),L=n("8c4f"),P=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"home-container"}},[t._m(0),n("div",{attrs:{id:"navigation-square-container"}},t._l(t.navSquareList,(function(e){return n("v-sheet",{directives:[{name:"ripple",rawName:"v-ripple"}],key:e.title,staticClass:"navigation-square",attrs:{elevation:"2"},on:{click:function(n){return t.squareClickedHandler(e.title)}}},[n("img",{attrs:{src:e.img,alt:"image"}}),n("h2",[t._v(t._s(e.title))])])})),1)])},O=[function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"landing-text"}},[n("h1",[t._v("Hi. I'm Nico.")]),n("p",[t._v("I'm a Computer Science student with a deep passion for software engineering. Checkout some of my work below and feel free to get in touch")])])}],E={name:"Home",mounted:function(){var t=this;return Object(o["a"])(regeneratorRuntime.mark((function e(){return regeneratorRuntime.wrap((function(e){while(1)switch(e.prev=e.next){case 0:return e.next=2,t.checkAuth();case 2:case"end":return e.stop()}}),e)})))()},props:{msg:String},methods:{checkAuth:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(){var e,n;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return t.next=2,fetch(c+"/checkAuth");case 2:if(e=t.sent,!e.ok){t.next=9;break}return t.next=6,e.json();case 6:n=t.sent,this.isUserLoggedIn=n.isLoggedIn,this.$root.$emit("authCheck",n);case 9:case"end":return t.stop()}}),t,this)})));function e(){return t.apply(this,arguments)}return e}(),squareClickedHandler:function(t){this.$root.$emit("pchange",t),this.$router.push({name:t})}},data:function(){return{navSquareList:[{title:"Resume",img:"/assets/Resume.svg"},{title:"Projects",img:"/assets/Projects.svg"},{title:"About",img:"/assets/About.svg"},{title:"Connect",img:"/assets/Connect.svg"}]}}},D=E,R=(n("be6f"),Object(p["a"])(D,P,O,!1,null,"75a8255a",null)),A=R.exports,I=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"resume-container"}},[n("v-tabs",{staticStyle:{width:"100%"},attrs:{"background-color":"transparent",grow:""},model:{value:t.tab,callback:function(e){t.tab=e},expression:"tab"}},t._l(t.items,(function(e){return n("v-tab",{key:e.tab},[t._v(" "+t._s(e.tab)+" ")])})),1),n("v-tabs-items",{attrs:{id:"resume-tab-item-container"},model:{value:t.tab,callback:function(e){t.tab=e},expression:"tab"}},[n("v-tab-item",{staticClass:"resume-tab-item"},[n("WebResume")],1),n("v-tab-item",{staticClass:"resume-tab-item"},[n("object",{attrs:{width:"100%",height:"90%",type:"application/pdf",data:"/nicolas-capparelli-resume.pdf"}},[n("div",{staticStyle:{height:"100%",width:"90%",display:"flex","flex-direction":"column","align-items":"center","justify-content":"center",margin:"auto"}},[n("p",{staticStyle:{"font-size":"20px","text-align":"center"}},[t._v("PDF Preview is not supported on mobile, press the button to download")]),n("v-btn",{staticStyle:{"margin-bottom":"1rem","background-color":"#424242"},on:{click:function(e){return t.downloadPDF()}}},[t._v(" Donwload PDF ")])],1)])])],1)],1)},T=[],N=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"web-resume-component-container"}},[n("Education"),n("v-divider"),n("Experience"),n("v-divider"),n("Projects"),n("v-divider"),n("Skills"),n("v-divider"),n("Awards")],1)},F=[],M=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"education-container"}},[n("h2",{staticClass:"section-title"},[t._v("EDUCATION")]),n("div",{staticClass:"education-box"},[n("v-avatar",{attrs:{tile:""}},[n("img",{attrs:{src:"/assets/KSU.svg",alt:"Kennesaw State University"}})]),t._m(0)],1)])},$=[function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",[n("h1",{staticStyle:{"font-size":"18px"}},[t._v("Kennesaw State University")]),n("span",{staticStyle:{"font-size":"16px"}},[t._v("BS Computer Science, May 2022")]),n("br"),n("span",{staticStyle:{"font-size":"14px",color:"gray"}},[t._v("3.5 GPA")]),n("ul",{staticStyle:{margin:"1rem 0"}},[n("li",[t._v("Programming Principles I & II")]),n("li",[t._v("Intro to Software Engineering")]),n("li",[t._v("Intro to Database Programming")]),n("li",[t._v("Computer Org & Arch")]),n("li",[t._v("Discrete Math")])])])}],q={name:"Education"},z=q,B=(n("a785"),Object(p["a"])(z,M,$,!1,null,"76e9f9e1",null)),H=B.exports,U=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"experience-container"}},[n("h2",{staticClass:"section-title"},[t._v("EXPERIENCE")]),n("div",{attrs:{id:"experience-list"}},t._l(t.experienceList,(function(e,i){return n("div",{key:e.dates,staticClass:"experience-box"},[n("v-avatar",[n("img",{attrs:{src:e.image,alt:e.employer}})]),n("div",{staticClass:"experience-box-content"},[n("h1",{staticStyle:{"font-size":"18px"}},[t._v(t._s(e.employer))]),n("span",{staticStyle:{"font-size":"16px"}},[t._v(t._s(e.title))]),n("br"),n("span",{staticStyle:{"font-size":"14px",color:"gray"}},[t._v(t._s(e.dates))]),n("ul",{staticStyle:{margin:"1rem 0"}},t._l(e.bullets,(function(e,i){return n("li",{key:i,staticStyle:{margin:"0.5rem 0"}},[t._v(" "+t._s(e)+" ")])})),0),i<t.experienceList.length-1?n("v-divider"):t._e()],1)],1)})),0)])},J=[],G={name:"Experience",props:{msg:String},data:function(){return{experienceList:[{image:"/assets/google.svg",employer:"Google",title:"STEP Intern",dates:"May 2020 - August 2020",bullets:[]},{image:"/assets/stanford.svg",employer:"Stanford",title:"Volunteer Section Leader",dates:"April 2020 - May 2020",bullets:["Participated in Stanford's Covid-19 online programming course as a section leader","Taught programming principles to a group of 10 students on a weekly basis","Section topics included: variables, loops, functions, arrays, image manipulation, and data science"]},{image:"/assets/motorola-fixed.svg",employer:"Motorla Solutions",title:"Software Engineering Intern",dates:"May 2019 - August 2019",bullets:["Created fault alerts on radios that caught over 50 critical bugs and failures before reaching production","Developed Python scripts to automate the transfer of logs and statistics between 3 internal company databases","Worked across 4 teams to design and implement an error logging database using Azure Cloud and ElasticSearch","Created and executed over 60 unit tests to detect flaws in proprietary software modules"]},{image:"/assets/contractor.svg",employer:"Independent Software Contractor",title:"Software Developer",dates:"November 2016 - May 2019",bullets:["Created and utilized an incremental project management model to maintain constant contact with clients and their needs","Designed and implemented a system to synchronize online and offline user data across multiple devices","Designed and implemented a central hub for commanding and monitoring social media advertising bots","Implemented a multi-tier recurring payment system using Stripe"]},{image:"/assets/fulton.png",employer:"Fulton County Schools",title:"Network Engineer Shadow",dates:"May 2017 - July 2017",bullets:["Worked alongside professional network engineers to set up and maintain physical network servers across 15 schools","Monitored and recorded network performance across all Fulton schools using enterprise IT Management Software (Solarwinds)","Participated in large scale deployment of factory images for over 1,000 devices"]}]}}},W=G,K=(n("9a07"),Object(p["a"])(W,U,J,!1,null,"8541db0a",null)),V=K.exports,X=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",[n("h2",{staticClass:"section-title"},[t._v("PROJECTS")]),n("v-btn",{staticStyle:{"margin-bottom":"1rem","background-color":"#424242"},on:{click:function(e){return t.goToProjects()}}},[t._v(" Projects Page ")])],1)},Y=[],Q={name:"Projects",props:{msg:String},methods:{goToProjects:function(){this.$router.push({name:"Projects"})}}},Z=Q,tt=(n("85fe"),Object(p["a"])(Z,X,Y,!1,null,"321173e6",null)),et=tt.exports,nt=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"skills-container"}},[n("h2",{staticClass:"section-title"},[t._v("SKILLS")]),n("div",{attrs:{id:"skills-list"}},[n("div",{staticClass:"skill-chip-outer"},[n("span",{staticStyle:{color:"#DBDBDB","font-weight":"bold","font-size":"18px"}},[t._v("Languages")]),n("div",{staticClass:"skill-chip-inner"},t._l(t.languageList,(function(e){return n("span",{key:e,staticClass:"chip"},[t._v(" "+t._s(e)+" ")])})),0)]),n("div",{staticClass:"skill-chip-outer"},[n("span",{staticStyle:{color:"#DBDBDB","font-weight":"bold","font-size":"18px"}},[t._v("Technologies & Frameworks")]),n("div",{staticClass:"skill-chip-inner"},t._l(t.technologyList,(function(e){return n("span",{key:e,staticClass:"chip"},[t._v(" "+t._s(e)+" ")])})),0)]),n("div",{staticClass:"skill-chip-outer"},[n("span",{staticStyle:{color:"#DBDBDB","font-weight":"bold","font-size":"18px"}},[t._v("Skills")]),n("div",{staticClass:"skill-chip-inner",attrs:{id:"skills-inner"}},t._l(t.skillList,(function(e){return n("span",{key:e,staticClass:"chip"},[t._v(" "+t._s(e)+" ")])})),0)])])])},it=[],at={name:"Template",props:{msg:String},data:function(){return{languageList:["Javascript","Python","Java","Swift","Kotlin","C#"],technologyList:["Git","Github","Bitbucket","GIRA","VueJS","ReactJS","NodeJS","Android","IOS","MongoDB","Elastic","Azure","Google Cloud"],skillList:["Software Architecture","UI/UX Design","Unit Testing","Web Dev","App Dev","Backend Dev"]}}},rt=at,st=(n("e422"),Object(p["a"])(rt,nt,it,!1,null,"02e87bd9",null)),ot=st.exports,ct=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"awards-container"}},[n("h2",{staticClass:"section-title"},[t._v("AWARDS")]),n("div",{attrs:{id:"award-list"}},t._l(t.awardList,(function(e,i){return n("div",{key:i,staticClass:"award-box"},[n("v-avatar",{attrs:{tile:""}},[n("img",{attrs:{src:e.image,alt:e.giver}})]),n("div",[n("h1",{staticStyle:{"font-size":"18px"}},[t._v(t._s(e.title))]),n("span",{staticStyle:{"font-size":"16px"}},[t._v(t._s(e.giver))]),n("br"),n("span",{staticStyle:{"font-size":"14px",color:"gray"}},[t._v(t._s(e.date))]),n("p",{staticStyle:{margin:"1rem 0"}},[t._v(t._s(e.description))]),i<t.awardList.length-1?n("v-divider"):t._e()],1)],1)})),0)])},lt=[],ut={name:"Awards",data:function(){return{awardList:[{image:"/assets/trophy.svg",title:"1st Place Hackathon Winner ",giver:"Kennesaw State Universtiy | Equifax",date:"September 2019",description:"Placed 1st on Equifax's security challenge by applying knowledge of multiprocessing & multithreading in Python and C++ to optimize a password cracking algorithm"},{image:"/assets/scholar.svg",title:"HSF Scholar",giver:"Hispanic Scholarship Fund",date:"September 2019",description:'"One of approximately 10,000 of the best and brightest Hispanic students in the country who are eligible to receive HSF\'s Scholar Support Services"'},{image:"/assets/scholar.svg",title:"HOPE Scholarship",giver:"State of Georgia",date:"May 2018",description:'"The HOPE Scholarship program is for students who have demonstrated academic achievement and are seeking a college degree."'}]}}},mt=ut,pt=(n("7fca"),Object(p["a"])(mt,ct,lt,!1,null,"3decf580",null)),dt=pt.exports,ht={name:"Template",props:{msg:String},components:{Education:H,Experience:V,Skills:ot,Projects:et,Awards:dt}},gt=ht,vt=(n("6258"),Object(p["a"])(gt,N,F,!1,null,"4bba561f",null)),ft=vt.exports,bt={name:"Template",components:{WebResume:ft},props:{msg:String},methods:{downloadPDF:function(){window.open("/assets/resume.pdf")}},data:function(){return{tab:null,items:[{tab:"WEB"},{tab:"PDF"}]}}},yt=bt,wt=(n("05c0"),Object(p["a"])(yt,I,T,!1,null,"06852a12",null)),_t=wt.exports,kt=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"project-container"}},[n("v-combobox",{staticStyle:{width:"100%","margin-top":"1rem"},attrs:{"item-color":"#999999",color:"#FFFFFF",items:t.items,chips:"",clearable:"",label:"Filter by language, technology...",multiple:"","single-line":"",solo:""},on:{input:t.filterProjectCards},scopedSlots:t._u([{key:"selection",fn:function(e){var i=e.attrs,a=e.item,r=e.select,s=e.selected,o=e.index;return[o<=t.numChipsDisplayed?n("v-chip",t._b({attrs:{"input-value":s,close:""},on:{click:r,"click:close":function(e){return t.remove(a)}}},"v-chip",i,!1),[n("strong",[t._v(t._s(a))])]):t._e(),o===t.numChipsDisplayed+1?n("span",{staticClass:"grey--text caption"},[t._v(" (+"+t._s(t.chips.length-1)+" others) ")]):t._e()]}}]),model:{value:t.chips,callback:function(e){t.chips=e},expression:"chips"}}),n("div",{attrs:{id:"project-list"}},t._l(t.visibleProjectList,(function(t){return n("ProjectCard",{key:t.title,attrs:{title:t.title,description:t.description,githubLink:t.githubLink,webLink:t.webLink,technologyList:t.technologyList,imageList:t.imageList}})})),1)],1)},xt=[],Ct=(n("c975"),n("a434"),n("b85c")),St=n("2909"),jt=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"project-card-container"},[n("div",{staticClass:"project-content"},[n("div",{staticClass:"project-header"},[n("h1",{staticStyle:{"font-size":"22px"}},[t._v(t._s(t.title))]),n("p",[t._v(t._s(t.description))])]),n("div",{staticClass:"project-footer"},[n("div",{staticClass:"project-icon-container"},[t.webLink?n("v-btn",{attrs:{icon:"",small:""},on:{click:function(e){return t.openWebLink()}}},[n("v-icon",[t._v("mdi-web")])],1):t._e(),t.githubLink?n("a",{directives:[{name:"ripple",rawName:"v-ripple"}],staticStyle:{height:"28px",width:"28px","border-radius":"28px"},attrs:{href:t.githubLink}},[n("img",{staticStyle:{height:"28px",width:"28px"},attrs:{src:"/assets/GitHub.svg",alt:"Github"}})]):t._e()],1),n("v-divider",{staticStyle:{height:"1px",width:"100%","background-color":"gray",margin:"0.5rem 0 0.7rem 0"}}),n("div",{staticClass:"project-chip-container"},t._l(t.technologyList,(function(e){return n("span",{key:e,staticClass:"chip"},[t._v(" "+t._s(e)+" ")])})),0)],1)]),t.imageList.length>0?n("div",{staticClass:"project-expansion-container"},[t.isExpanded?n("div",{staticStyle:{height:"350px",width:"100%",padding:"16px"}},[n("v-carousel",{attrs:{height:"100%","show-arrows":!1}},t._l(t.imageList,(function(t,e){return n("v-carousel-item",{key:e,attrs:{src:t}})})),1)],1):t._e(),n("v-btn",{attrs:{icon:"",large:""},on:{click:function(e){return t.expandCard()}}},[n("v-icon",[t._v(t._s(t.arrowIcon))])],1)],1):t._e()])},Lt=[],Pt={name:"ProjectCard",props:{title:String,description:String,githubLink:String,webLink:String,technologyList:Array,imageList:Array},data:function(){return{isExpanded:!1,arrowIcon:"mdi-chevron-down"}},methods:{expandCard:function(){this.isExpanded=!this.isExpanded,this.arrowIcon=this.isExpanded?"mdi-chevron-up":"mdi-chevron-down"},openWebLink:function(){window.open(this.webLink)}}},Ot=Pt,Et=(n("ad51"),Object(p["a"])(Ot,jt,Lt,!1,null,"05041368",null)),Dt=Et.exports,Rt=[{title:"Portfolio Website",description:"A simple website to display my information, resume, and projects as well as provide contact information.",githubLink:"https://github.com/NicolasCapparelli/step-indv/tree/master/portfolio",webLink:"capparelli-step-2020.appspot.com",technologyList:["VueJS","HTML","CSS","Javascript"],imageList:[]},{title:"Instructor Toolkit",description:"Cloud based administration tool & management service aiming to support over 500 basic life support training centers and their instructors (work in progress)",githubLink:null,webLink:"https://instructortk.com",technologyList:["React","NodesJS","HTML","CSS","Javascript","Swift"],imageList:[]},{title:"Project Valery",description:"A personal desktop voice assistant capable of controlling the computer, searching the web, and opening files in their respective applications",githubLink:"https://github.com/NicolasCapparelli/Project-Valery",webLink:null,technologyList:["Python","speech_recognition"],imageList:["https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg","https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg","https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg"]},{title:"NC9v2",description:"A chat bot on the discord platform used for retrieving financial data through the iex stock Exchange api. At its peak, this bot served over 1,000 users across 5 different communities",githubLink:"https://github.com/NicolasCapparelli/NC9v2",webLink:null,technologyList:["Python"],imageList:["https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg","https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg","https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg"]},{title:"RemindMe",description:"A lightweight python module to create and notify users of reminders using asynchronous functions",githubLink:"https://github.com/NicolasCapparelli/NC9v2/blob/master/remindme.py",webLink:null,technologyList:["Python"],imageList:["https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg","https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg","https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg"]},{title:"Clip History",description:"A simple and lightweight clipboard history tool for windows",githubLink:"https://github.com/NicolasCapparelli/Clip-History",webLink:null,technologyList:["C#"],imageList:["https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg","https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg","https://cdn.vuetifyjs.com/images/carousel/squirrel.jpg"]}],At={name:"Projects",mounted:function(){window.scrollTo(0,0)},data:function(){return{numChipsDisplayed:0,chips:[],items:["Javascript","Python","C#","VueJS","React","HTML","CSS"],visibleProjectList:Rt}},components:{ProjectCard:Dt},methods:{remove:function(t){this.chips.splice(this.chips.indexOf(t),1),this.chips=Object(St["a"])(this.chips)},filterProjectCards:function(t){if(t.length<1)this.visibleProjectList=Rt;else{var e,n=[],i=Object(Ct["a"])(Rt);try{for(i.s();!(e=i.n()).done;){var a,r=e.value,s=Object(Ct["a"])(r.technologyList);try{for(s.s();!(a=s.n()).done;){var o,c=a.value,l=Object(Ct["a"])(t);try{for(l.s();!(o=l.n()).done;){var u=o.value;if(u.toLowerCase()===c.toLowerCase()){n.push(r);break}}}catch(m){l.e(m)}finally{l.f()}}}catch(m){s.e(m)}finally{s.f()}}}catch(m){i.e(m)}finally{i.f()}this.visibleProjectList=n}}}},It=At,Tt=(n("a307"),Object(p["a"])(It,kt,xt,!1,null,"e7d18eca",null)),Nt=Tt.exports,Ft=function(){var t=this,e=t.$createElement;t._self._c;return t._m(0)},Mt=[function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"about-container"}},[n("div",{staticClass:"about-section"},[n("h1",{staticClass:"section-title"},[t._v("Info")]),n("ul",[n("li",[t._v("🇺🇸 U.S. Citizen")]),n("li",[t._v("20 Years Old")]),n("li",[t._v("Class of 2022")]),n("li",[t._v("Pursuing B.S. in Computer Science")]),n("li",[t._v("Kennesaw State University")])])]),n("div",{staticClass:"about-section"},[n("h1",{staticClass:"section-title"},[t._v("About")]),n("p",{staticStyle:{"margin-top":"0.5rem"}},[t._v("Ever since I was 15, I've loved the process of turning ideas into reality; so naturally, I gained great interest in Software Development and Computer Science.")]),n("p",{staticStyle:{"margin-top":"0.5rem"}},[t._v("Since then, I've worked on several personal projects and freelance work in a variety of languages & frameworks to expand my knowledge and gather experience.")]),n("p",{staticStyle:{"margin-top":"0.5rem"}},[t._v("From native mobile apps to voice assistants, I have a cast a wide net in order to diversify my skillset and tackle new challenges. Most recently I have started taking part in internships to work with large teams and learn from those around me.")])]),n("div",{staticClass:"about-section"},[n("h1",{staticClass:"section-title"},[t._v("Interests")]),n("ul",[n("li",[t._v("Software Engineering")]),n("li",[t._v("Finance")]),n("li",[t._v("UFC")]),n("li",[t._v("Football")])])]),n("div",{staticClass:"about-section"},[n("h1",{staticClass:"section-title"},[t._v("Location")]),n("p",[t._v("I currently reside in Georgia, United States")]),n("div",{attrs:{id:"globe"}})])])}],$t="AIzaSyCvEL807xCAiPpRFTM2C2tO1mWD-vpvXQI",qt="gmapsCallback",zt=!!window.google,Bt=new Promise((function(t,e){x=t,C=e}));function Ht(){if(zt)return Bt;zt=!0,window[qt]=function(){return x(window.google)};var t=document.createElement("script");return t.async=!0,t.defer=!0,t.src="https://maps.googleapis.com/maps/api/js?key=".concat($t,"&callback=").concat(qt),t.onerror=C,document.querySelector("head").appendChild(t),Bt}var Ut={name:"About",mounted:function(){return Object(o["a"])(regeneratorRuntime.mark((function t(){var e,n,i;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,Ht();case 3:e=t.sent,n=new e.maps.Geocoder,i=new e.maps.Map(document.getElementById("globe")),n.geocode({address:"Georgia"},(function(t,n){if("OK"!==n||!t[0])throw new Error(n);var a=new e.maps.Marker({position:t[0].geometry.location,title:"Hey there!"});i.setCenter(t[0].geometry.location),i.fitBounds(t[0].geometry.viewport),a.setMap(i)})),t.next=12;break;case 9:t.prev=9,t.t0=t["catch"](0),console.error(t.t0);case 12:case"end":return t.stop()}}),t,null,[[0,9]])})))()},props:{msg:String}},Jt=Ut,Gt=(n("a477"),Object(p["a"])(Jt,Ft,Mt,!1,null,"0ddd878b",null)),Wt=Gt.exports,Kt=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"connect-container"}},[n("v-alert",{attrs:{id:"connect-alert",type:t.alertType,dismissible:"",elevation:"3",transition:"scroll-y-transition"},model:{value:t.alert,callback:function(e){t.alert=e},expression:"alert"}},[t._v(" "+t._s(t.alertMessage)+" ")]),n("h1",{staticStyle:{width:"100%","margin-top":"1rem"}},[t._v("Connect")]),t._m(0),n("h1",{staticStyle:{width:"100%"}},[t._v("Leave a Comment")]),n("div",{attrs:{id:"comments-container"}},[n("div",{staticClass:"comment",attrs:{id:"message-input-container"}},[n("v-text-field",{staticStyle:{width:"100%"},attrs:{label:"Name",color:t.colors.textPrimary,dark:""},model:{value:t.name,callback:function(e){t.name=e},expression:"name"}}),n("v-textarea",{staticStyle:{width:"100%"},attrs:{label:"Message",color:t.colors.textPrimary,dark:""},model:{value:t.message,callback:function(e){t.message=e},expression:"message"}}),t.isUserLoggedIn?n("v-btn",{staticStyle:{position:"relative",margin:"0.5rem 0 0 auto"},attrs:{text:"",disabled:t.isProcessingNewComment},on:{click:function(e){return t.sendComment()}}},[t._v(" Send "),t.isProcessingNewComment?n("v-progress-circular",{staticStyle:{position:"absolute","background-color":"#555555",width:"100%"},attrs:{indeterminate:"",color:"primary"}}):t._e()],1):n("v-btn",{staticStyle:{position:"relative",margin:"0.5rem 0 0 auto"},attrs:{text:""},on:{click:function(e){return t.authenticate()}}},[t._v(" Login to comment ")])],1),n("div",{attrs:{id:"comment-actions-container"}},[n("v-select",{staticStyle:{"max-width":"100px","margin-top":"1rem"},style:{justifyContent:t.justifyActionBox},attrs:{items:t.numCommentOptions,color:t.colors.textPrimary,label:"Number of Comments","prepend-icon":"filter_list","single-line":""},on:{change:function(e){return t.changeNumComments()}},model:{value:t.numComments,callback:function(e){t.numComments=e},expression:"numComments"}}),n("div",{attrs:{id:"pagination-container"}},[n("v-btn",{attrs:{icon:""},on:{click:function(e){return t.changeCommentPage(!1)}}},[n("v-icon",[t._v("mdi-chevron-left")])],1),n("span",{staticStyle:{margin:"0 0.5rem"}},[t._v(t._s(t.page))]),n("v-btn",{attrs:{icon:""},on:{click:function(e){return t.changeCommentPage(!0)}}},[n("v-icon",[t._v("mdi-chevron-right")])],1)],1),t.possibleAdmin?n("v-btn",{attrs:{icon:"",color:t.colors.textPrimary},on:{click:function(e){return t.deleteComments()}}},[n("v-icon",[t._v("mdi-close")])],1):t._e()],1),t.firstLoadDone?n("div",{attrs:{id:"comment-list"}},t._l(t.commentList,(function(e){return n("Comment",{key:e.id,attrs:{commentData:e,successfulChangeCallback:t.commentSuccessCallback}})})),1):t._e(),t.firstLoadDone?t._e():n("div",{attrs:{id:"skeleton-load-container"}},t._l(t.numComments,(function(t){return n("v-skeleton-loader",{key:t,staticStyle:{width:"100%","margin-bottom":"1rem"},attrs:{color:"#555555",type:"list-item-three-line"}})})),1)])],1)},Vt=[function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"social-media-container"}},[n("img",{staticClass:"media-box",attrs:{src:"/assets/GitHub.svg",alt:"linkedin"}}),n("img",{staticClass:"media-box",attrs:{src:"/assets/GitHub.svg",alt:"linkedin"}}),n("img",{staticClass:"media-box",attrs:{src:"/assets/GitHub.svg",alt:"linkedin"}})])}],Xt=(n("b0c0"),n("a9e3"),n("aff5"),n("ac1f"),n("25f0"),n("1276"),function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"comment"},[n("div",{staticClass:"comment-header-container"},[n("h1",{staticStyle:{"font-size":"18px"}},[t._v(t._s(t.commentData.name))]),n("span",{staticStyle:{color:"#a3a2a2","font-size":"14px"}},[t._v(t._s(t.commentData.timestamp))])]),n("p",{staticStyle:{width:"100%"}},[t._v(t._s(t.commentData.message))]),n("div",{staticStyle:{width:"100%",display:"flex","align-items":"center","justify-content":"flex-end"}},[n("div",{staticStyle:{"margin-right":"1.5rem"}},[n("v-btn",{staticClass:"vote-button",attrs:{icon:"",small:""},on:{click:function(e){return t.commentRatingChange(!0)}}},[n("v-icon",[t._v("mdi-thumb-up")])],1),n("span",[t._v(t._s(t.commentData.upvotes))])],1),n("div",[n("v-btn",{staticClass:"vote-button",attrs:{icon:"",small:""},on:{click:function(e){return t.commentRatingChange(!1)}}},[n("v-icon",[t._v("mdi-thumb-down")])],1),n("span",[t._v(t._s(t.commentData.downvotes))])],1)])])}),Yt=[],Qt={name:"Comment",props:{commentData:Object,successfulChangeCallback:Function},methods:{commentRatingChange:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(e){var n;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return t.next=2,fetch(c+"/rating",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({commentID:this.commentData.id,isUpvote:e})});case 2:n=t.sent,n.ok?this.successfulChangeCallback():alert("There was an error processing your request, please try again later");case 4:case"end":return t.stop()}}),t,this)})));function e(e){return t.apply(this,arguments)}return e}()}},Zt=Qt,te=(n("72b3"),Object(p["a"])(Zt,Xt,Yt,!1,null,"1498e566",null)),ee=te.exports,ne={name:"Connect",created:function(){var t=this;return Object(o["a"])(regeneratorRuntime.mark((function e(){var n,i;return regeneratorRuntime.wrap((function(e){while(1)switch(e.prev=e.next){case 0:return e.next=2,t.getDataFromServer();case 2:n=e.sent,t.updateCommentList(n),t.firstLoadDone=!0,i=ie("xcvd_admin_key_xcvi"),i.length>0&&(t.possibleAdmin=!0,t.adminKey=i);case 7:case"end":return e.stop()}}),e)})))()},mounted:function(){var t=this;return Object(o["a"])(regeneratorRuntime.mark((function e(){return regeneratorRuntime.wrap((function(e){while(1)switch(e.prev=e.next){case 0:return e.next=2,t.checkAuth();case 2:case"end":return e.stop()}}),e)})))()},data:function(){return{name:"",message:"",isUserLoggedIn:!1,page:1,lastPage:Number.MAX_SAFE_INTEGER,numComments:5,commentList:[],numCommentOptions:[5,10,15],alertMessage:"",alertType:"success",alert:!1,firstLoadDone:!1,isProcessingNewComment:!1,possibleAdmin:!1,justifyActionBox:"flex-start",colors:l}},components:{Comment:ee},methods:{checkAuth:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(){var e,n;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return t.next=2,fetch(c+"/checkAuth");case 2:if(e=t.sent,!e.ok){t.next=9;break}return t.next=6,e.json();case 6:n=t.sent,this.isUserLoggedIn=n.isLoggedIn,this.$root.$emit("authCheck",n);case 9:case"end":return t.stop()}}),t,this)})));function e(){return t.apply(this,arguments)}return e}(),authenticate:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(){var e,n;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return t.next=2,fetch(c+"/authenticate");case 2:if(e=t.sent,!e.ok){t.next=10;break}return t.next=6,e.json();case 6:n=t.sent,location.href=n.authURL,t.next=11;break;case 10:alert("There was an issue processing your request, please try again later");case 11:case"end":return t.stop()}}),t)})));function e(){return t.apply(this,arguments)}return e}(),getDataFromServer:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(){var e,n;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return t.next=2,fetch(c+"/comments?numComments="+this.numComments+"&page="+this.page);case 2:if(e=t.sent,!e.ok){t.next=9;break}return t.next=6,e.json();case 6:n=t.sent,t.next=11;break;case 9:alert("There appears to be a connection error, please try again later"),n=[];case 11:return this.isProcessingNewComment=!1,t.abrupt("return",n);case 13:case"end":return t.stop()}}),t,this)})));function e(){return t.apply(this,arguments)}return e}(),updateCommentList:function(t){var e,n=Object(Ct["a"])(t);try{for(n.s();!(e=n.n()).done;){var i=e.value,a=new Date(i.timestamp);i.timestamp=(a.getMonth()+1).toString()+"/"+a.getDay().toString()+"/"+a.getFullYear().toString()}}catch(r){n.e(r)}finally{n.f()}this.commentList=t},sendComment:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(){var e,n,i,a;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return this.isProcessingNewComment=!0,e=this.name,n=this.message,t.next=5,fetch(c+"/comments",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({datePosted:null,name:e,message:n})});case 5:return i=t.sent,i.ok?(this.name="",this.message=""):alert("There appears to be a connection error, please try again later"),this.isProcessingNewComment=!1,t.next=10,this.getDataFromServer();case 10:a=t.sent,this.updateCommentList(a);case 12:case"end":return t.stop()}}),t,this)})));function e(){return t.apply(this,arguments)}return e}(),deleteComments:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(){var e,n;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return t.next=2,fetch(c+"/delete-data",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({adminKey:this.adminKey})});case 2:if(e=t.sent,!e.ok){t.next=12;break}return this.alertType="success",this.alertMessage="Comments delete successfully",t.next=8,this.getDataFromServer();case 8:n=t.sent,this.updateCommentList(n),t.next=14;break;case 12:this.alertType="error",this.alertMessage="You are not in the sudoers file, this incident will be reported";case 14:this.alert=!0;case 15:case"end":return t.stop()}}),t,this)})));function e(){return t.apply(this,arguments)}return e}(),commentSuccessCallback:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(){var e;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return t.next=2,this.getDataFromServer();case 2:e=t.sent,console.log(e),this.updateCommentList(e);case 5:case"end":return t.stop()}}),t,this)})));function e(){return t.apply(this,arguments)}return e}(),changeNumComments:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(){var e,n;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return this.lastPage=Number.MAX_SAFE_INTEGER,t.next=3,this.getDataFromServer();case 3:if(e=t.sent,!(e.length<1)){t.next=12;break}return this.lastPage=this.page-1,1!=this.page&&(this.page-=1),t.next=9,this.getDataFromServer();case 9:return n=t.sent,this.updateCommentList(n),t.abrupt("return");case 12:this.updateCommentList(e);case 13:case"end":return t.stop()}}),t,this)})));function e(){return t.apply(this,arguments)}return e}(),changeCommentPage:function(){var t=Object(o["a"])(regeneratorRuntime.mark((function t(e){var n;return regeneratorRuntime.wrap((function(t){while(1)switch(t.prev=t.next){case 0:return e?this.page<this.lastPage&&(this.page+=1):this.page>1&&(this.page-=1),t.next=3,this.getDataFromServer();case 3:if(n=t.sent,!(n.length<1)){t.next=10;break}return this.lastPage=this.page-1,1!=this.page&&(this.page-=1),t.abrupt("return");case 10:this.updateCommentList(n);case 11:case"end":return t.stop()}}),t,this)})));function e(e){return t.apply(this,arguments)}return e}()}};function ie(t){for(var e=t+"=",n=document.cookie,i=n.split(";"),a=0;a<i.length;a++){var r=i[a];while(" "==r.charAt(0))r=r.substring(1);if(0==r.indexOf(e))return r.substring(e.length,r.length)}return""}var ae=ne,re=(n("52a4"),Object(p["a"])(ae,Kt,Vt,!1,null,null,null)),se=re.exports;y["default"].config.productionTip=!1,y["default"].use(L["a"]),y["default"].use(w["a"]);var oe=[{name:"Home",path:"/",component:A},{name:"Resume",path:"/resume",component:_t},{name:"Projects",path:"/projects",component:Nt},{name:"About",path:"/about",component:Wt},{name:"Connect",path:"/connect",component:se}],ce=new L["a"]({routes:oe});new y["default"]({vuetify:j,router:ce,render:function(t){return t(b)}}).$mount("#app")},6258:function(t,e,n){"use strict";var i=n("1395"),a=n.n(i);a.a},"659c":function(t,e,n){},"72b3":function(t,e,n){"use strict";var i=n("cf4b"),a=n.n(i);a.a},"7fca":function(t,e,n){"use strict";var i=n("94d0"),a=n.n(i);a.a},"85ec":function(t,e,n){},"85fe":function(t,e,n){"use strict";var i=n("15f2"),a=n.n(i);a.a},8784:function(t,e,n){},"90dc":function(t,e,n){},"94d0":function(t,e,n){},"9a07":function(t,e,n){"use strict";var i=n("659c"),a=n.n(i);a.a},a307:function(t,e,n){"use strict";var i=n("4276"),a=n.n(i);a.a},a477:function(t,e,n){"use strict";var i=n("c94d"),a=n.n(i);a.a},a785:function(t,e,n){"use strict";var i=n("1a74"),a=n.n(i);a.a},a845:function(t,e,n){},ad51:function(t,e,n){"use strict";var i=n("8784"),a=n.n(i);a.a},b2d5:function(t,e,n){},be6f:function(t,e,n){"use strict";var i=n("90dc"),a=n.n(i);a.a},c84a:function(t,e,n){},c94d:function(t,e,n){},cf4b:function(t,e,n){},e422:function(t,e,n){"use strict";var i=n("c84a"),a=n.n(i);a.a}});
-//# sourceMappingURL=app.92da4305.js.map
