+++ portfolio/src/utils/constants.js
@@ -1,6 +1,6 @@
 
-const WEBSITE_URL = "https://8080-b83fc153-d2cf-481d-a321-9342cdf80f21.us-east1.cloudshell.dev"
-// const WEBSITE_URL = "http://capparelli-step-2020.appspot.com"
+//const WEBSITE_URL = "https://8080-b83fc153-d2cf-481d-a321-9342cdf80f21.us-east1.cloudshell.dev"
+const WEBSITE_URL = "http://capparelli-step-2020.appspot.com"
 
 const COLORS = {
     textPrimary: "#DBDBDB"
